improve testing coverage of akka-http, extract Person class
