+++ src/test/scala/org/zalando/jsonapi/json/ExampleSpec.scala
@@ -12,18 +12,6 @@ class ExampleSpec extends WordSpec with MustMatchers with SprayJsonJsonapiProtoc
   "JsonapiRootObject" when {
     "using root object serializer" must {
       "serialize accordingly" in {
-        case class Person(id: Int, name: String)
-
-        implicit val personJsonapiRootObjectWriter: JsonapiRootObjectWriter[Person] = new JsonapiRootObjectWriter[Person] {
-          override def toJsonapi(person: Person) = {
-            RootObject(data = Some(ResourceObject(
-              `type` = "person",
-              id = Some(person.id.toString),
-              attributes = Some(List(
-                Attribute("name", StringValue(person.name))
-              )), links = None)))
-          }
-        }
 
         val json =
           """
@@ -43,8 +31,6 @@ class ExampleSpec extends WordSpec with MustMatchers with SprayJsonJsonapiProtoc
     }
 
     "serialize accordingly with links object in data array" in {
-      case class Person(id: Int, name: String)
-
       implicit val personJsonapiRootObjectWriter: JsonapiRootObjectWriter[Person] = new JsonapiRootObjectWriter[Person] {
         override def toJsonapi(person: Person) = {
           RootObject(data = Some(ResourceObject(
@@ -76,8 +62,6 @@ class ExampleSpec extends WordSpec with MustMatchers with SprayJsonJsonapiProtoc
     }
 
     "serialize accordingly with links object in root object " in {
-      case class Person(id: Int, name: String)
-
       implicit val personJsonapiRootObjectWriter: JsonapiRootObjectWriter[Person] = new JsonapiRootObjectWriter[Person] {
         override def toJsonapi(person: Person) = {
           RootObject(data = Some(ResourceObject(
